+++ CHANGELOG.md
@@ -2,9 +2,15 @@
 
 ### Added
 
+- "Show More" Button inside Units to avoid being punched in the face with too many exercises at once
+- Footer to exported PDF with centered "Made with Neverlose" text
+- Added "Open file" hyperlink to the export-finished dialog
+- Collapsible Training Units with toggle arrow to show/hide content
+
 ### Changed
+
+- Moved the "+" and "Load" button for Units to the left
 
 ### Fixed
 
 ### Removed
-